requirements.txt
Pillow==8.3.1
pyscreenshot==3.0
pywin32==301